added all tests in auth.test.js
